moved login/logout
