+++ frontend/src/App.tsx
@@ -9,10 +9,10 @@ import { createMuiTheme, makeStyles, ThemeProvider } from '@material-ui/core';
 import { useSwipeable } from 'react-swipeable';
 import axios from 'axios';
 import { useQuery } from 'react-query';
-import Login from './views/Login';
+import Login from './views/Auth/Login';
 import PlaylistView from './views/PlaylistView';
 import Track from './views/TrackPage';
-import Logout from './views/Logout';
+import Logout from './views/Auth/Logout';
 import Recent from './views/RecentPage';
 import Settings from './views/SettingsPage';
 import Top from './views/TopPage';
